Añadir alert de confirmar borrado de pedido

+++ src/Components/OrderItem.js
@@ -6,7 +6,7 @@ import { Button } from 'react-bootstrap';
 import axios from 'axios';
 import OrdersContext from '../Storage/OrdersContext';
 
-const OrderItem = ({ orderId, name, products }) => {
+const OrderItem = ({ orderId, name, products, date }) => {
 
     //TODO: Eliminar order
     // Delete order from context and database
@@ -14,20 +14,25 @@ const OrderItem = ({ orderId, name, products }) => {
     const ordersContext = useContext(OrdersContext);
 
     const handleDeleteOrder = () => {
-        console.log("Delete order " + orderId)
-        axios.delete("https://climbcrafters-default-rtdb.europe-west1.firebasedatabase.app/users/" + loginContext.email + "/pedidos/" + orderId + ".json")
-            .then((response) => {
-                console.log(response)
-                ordersContext.setReload(!ordersContext.reload)
-            }).catch((error) => {
-                console.log(error)
-            })
+        if (window.confirm('Are you sure you want to delete the item?')) {
+            axios.delete("https://climbcrafters-default-rtdb.europe-west1.firebasedatabase.app/users/" + loginContext.email + "/pedidos/" + orderId + ".json")
+                .then((response) => {
+                    console.log(response)
+                    ordersContext.setReload(!ordersContext.reload)
+                }).catch((error) => {
+                    console.log(error)
+                })
+        } else {
+
+        }
+
     }
 
     return (
         <div className="order">
             <Link to={`/order-details/${orderId}`}>
                 <div>
+                    <p>{date} </p>
                     <p>{name} </p>
                     <p>{products}</p>
                 </div>
